:art: refactor(paginate): Move paginate.py to `app/utils`
Seems like a much better home for this functionality.

GTC-1865
