+++ tests_v2/unit/app/utils/paginate/test_pagination_links_info.py
@@ -2,7 +2,7 @@ from unittest.mock import AsyncMock
 
 import pytest
 
-from app.paginate.paginate import PaginationLinks, paginate_collection
+from app.utils.paginate import PaginationLinks, paginate_collection
 
 DONT_CARE: int = 1
